refactor(state): share schema version guard
