+++ src/state/openclaw-state-db-cache.ts
@@ -10,10 +10,8 @@ import { isSqliteCorruptionError } from "../infra/sqlite-transaction.js";
 import { isSqliteSchemaVersionError } from "../infra/sqlite-user-version.js";
 import { readOpenClawDatabaseQuarantine } from "./openclaw-quarantine-store.js";
 import type { OpenClawStateDatabase } from "./openclaw-state-db-contract.js";
-import {
-  assertSupportedSchemaVersion,
-  createOpenClawDatabaseVerificationError,
-} from "./openclaw-state-db-maintenance.js";
+import { createOpenClawDatabaseVerificationError } from "./openclaw-state-db-maintenance.js";
+import { assertSupportedStateSchemaVersion } from "./openclaw-state-db-schema-version.js";
 
 const cachedDatabases = new Map<string, OpenClawStateDatabase>();
 const cachedDataVersionStatements = new WeakMap<
@@ -155,7 +153,7 @@ function getOpenClawStateDatabaseRuntimeFailure(pathname: string): Error | undef
     }
     // data_version is the cheap external-commit trigger. Re-read user_version
     // only when another connection changed the file.
-    assertSupportedSchemaVersion(cached.db, resolvedPath);
+    assertSupportedStateSchemaVersion(cached.db, resolvedPath);
     cachedDataVersions.set(cached.db, dataVersion);
     return undefined;
   } catch (error) {
