+++ src/state/openclaw-state-db-schema-version.ts
@@ -1,0 +1,19 @@
+import type { DatabaseSync } from "node:sqlite";
+import {
+  createNewerSqliteSchemaVersionError,
+  readSqliteUserVersion,
+} from "../infra/sqlite-user-version.js";
+import { OPENCLAW_STATE_SCHEMA_VERSION } from "./openclaw-state-db-contract.js";
+
+export function assertSupportedStateSchemaVersion(db: DatabaseSync, pathname: string): number {
+  const userVersion = readSqliteUserVersion(db);
+  if (userVersion > OPENCLAW_STATE_SCHEMA_VERSION) {
+    throw createNewerSqliteSchemaVersionError(
+      "OpenClaw state database",
+      pathname,
+      userVersion,
+      OPENCLAW_STATE_SCHEMA_VERSION,
+    );
+  }
+  return userVersion;
+}
